+++ src/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './sdkDappUtils';
 export * from './sdkDappCore';
-export * from './getCallbackUrl';
+export * from './getCallbackRoute';
